use a overlay for Picker iOS instead of inline
